requirements.txt
PyMuPDF>=1.22.0
openai>=1.30.1
python-dotenv>=1.0.0
Jinja2>=3.1.6
streamlit>=1.45.1
